backend-function/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python 
# azure-monitor-opentelemetry 

azure-functions
azure-search-documents
azure-storage-blob
azure-core
openai
python-dotenv
PyPDF2
livereload
pycryptodome